@@ -4,28 +4,36 @@
   "private": true,
   "scripts": {
     "clean": "npx rimraf .docusaurus build node_modules",
+    "lintDocs": "prettier --write ../docs/**/*.md ../docs/**/*.mdx",
+    "docusaurus": "docusaurus",
     "start": "docusaurus start",
-    "build": "docusaurus build",
+    "build": "npm run lintDocs && docusaurus clear && docusaurus build",
     "swizzle": "docusaurus swizzle",
     "deploy": "docusaurus deploy",
-    "lintDocs": "prettier --write ../docs/**/*.md ../docs/**/*.mdx"
+    "clear": "docusaurus clear",
+    "serve": "docusaurus serve",
+    "write-translations": "docusaurus write-translations",
+    "write-heading-ids": "docusaurus write-heading-ids"
   },
   "dependencies": {
-    "@docusaurus/core": "^2.0.0-rc.1",
-    "@docusaurus/preset-classic": "^2.0.0-rc.1",
-    "bootstrap": "^5.0.2",
-    "classnames": "^2.2.6",
+    "@docusaurus/core": "2.1.0",
+    "@docusaurus/preset-classic": "2.1.0",
+    "@mdx-js/react": "^1.6.22",
+    "clsx": "^1.2.1",
     "odata-uri": "file:../tools/odataUri",
     "odata-url-editor": "file:../tools/urlEditor",
     "prettier": "^2.7.1",
-    "prism-react-renderer": "^1.2.1",
-    "react": "^16.8.4",
-    "react-bootstrap": "^1.6.1",
-    "react-dom": "^16.8.4"
+    "prism-react-renderer": "^1.3.5",
+    "react": "^17.0.2",
+    "react-bootstrap": "^2.5.0",
+    "react-dom": "^17.0.2"
+  },
+  "devDependencies": {
+    "@docusaurus/module-type-aliases": "2.1.0"
   },
   "browserslist": {
     "production": [
-      ">0.2%",
+      ">0.5%",
       "not dead",
       "not op_mini all"
     ],
@@ -34,5 +42,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "engines": {
+    "node": ">=16.14"
   }
 }
