@@ -1,13 +1,13 @@
-// Generated from c:\git\odata-rapid\tools\rsdl\rsdl-js\grammar\rsdl.g4 by ANTLR 4.8
+// Generated from c:\git\odata-rapid\tools\rsdl\rsdl-js\grammar\rsdl.g4 by ANTLR 4.9.2
 // jshint ignore: start
-var antlr4 = require('antlr4/index');
+import antlr4 from 'antlr4';
 
 
 
-var serializedATN = ["\u0003\u608b\ua72a\u8133\ub9ed\u417c\u3be7\u7786\u5964",
-    "\u0002/\u01b0\b\u0001\u0004\u0002\t\u0002\u0004\u0003\t\u0003\u0004",
-    "\u0004\t\u0004\u0004\u0005\t\u0005\u0004\u0006\t\u0006\u0004\u0007\t",
-    "\u0007\u0004\b\t\b\u0004\t\t\t\u0004\n\t\n\u0004\u000b\t\u000b\u0004",
+const serializedATN = ["\u0003\u608b\ua72a\u8133\ub9ed\u417c\u3be7\u7786",
+    "\u5964\u0002/\u01b0\b\u0001\u0004\u0002\t\u0002\u0004\u0003\t\u0003",
+    "\u0004\u0004\t\u0004\u0004\u0005\t\u0005\u0004\u0006\t\u0006\u0004\u0007",
+    "\t\u0007\u0004\b\t\b\u0004\t\t\t\u0004\n\t\n\u0004\u000b\t\u000b\u0004",
     "\f\t\f\u0004\r\t\r\u0004\u000e\t\u000e\u0004\u000f\t\u000f\u0004\u0010",
     "\t\u0010\u0004\u0011\t\u0011\u0004\u0012\t\u0012\u0004\u0013\t\u0013",
     "\u0004\u0014\t\u0014\u0004\u0015\t\u0015\u0004\u0016\t\u0016\u0004\u0017",
@@ -275,24 +275,49 @@ var serializedATN = ["\u0003\u608b\ua72a\u8133\ub9ed\u417c\u3be7\u7786\u5964",
     "\u0184\u018a\u0193\u0197\u019f\u01a3\u01ac\u0003\b\u0002\u0002"].join("");
 
 
-var atn = new antlr4.atn.ATNDeserializer().deserialize(serializedATN);
+const atn = new antlr4.atn.ATNDeserializer().deserialize(serializedATN);
 
-var decisionsToDFA = atn.decisionToState.map( function(ds, index) { return new antlr4.dfa.DFA(ds, index); });
+const decisionsToDFA = atn.decisionToState.map( (ds, index) => new antlr4.dfa.DFA(ds, index) );
 
-function rsdlLexer(input) {
-	antlr4.Lexer.call(this, input);
-    this._interp = new antlr4.atn.LexerATNSimulator(this, atn, decisionsToDFA, new antlr4.PredictionContextCache());
-    return this;
+export default class rsdlLexer extends antlr4.Lexer {
+
+    static grammarFileName = "rsdl.g4";
+    static channelNames = [ "DEFAULT_TOKEN_CHANNEL", "HIDDEN" ];
+	static modeNames = [ "DEFAULT_MODE" ];
+	static literalNames = [ null, "'namespace'", "'include'", "'as'", "'type'", 
+                         "'{'", "'}'", "'extends'", "':'", "'['", "']'", 
+                         "'Boolean'", "'Date'", "'DateTime'", "'Decimal'", 
+                         "'('", "','", "')'", "'Double'", "'Duration'", 
+                         "'Integer'", "'String'", "'TimeOfDay'", "'enum'", 
+                         "'flags'", "'typedef'", "'service'", "'true'", 
+                         "'false'", "'null'", "'.'", "'/'", "'abstract'", 
+                         "'action'", "'function'", "'key'", "'?'" ];
+	static symbolicNames = [ null, null, null, null, null, null, null, null, 
+                          null, null, null, null, null, null, null, null, 
+                          null, null, null, null, null, null, null, null, 
+                          null, null, null, null, null, null, null, null, 
+                          "ABSTRACT", "ACTION", "FUNCTION", "KEY", "NULLABLE", 
+                          "NUMBER", "STRING", "EDM", "ID", "QID", "TID", 
+                          "DOC_COMMENT", "LINE_COMMENT", "WS" ];
+	static ruleNames = [ "T__0", "T__1", "T__2", "T__3", "T__4", "T__5", "T__6", 
+                      "T__7", "T__8", "T__9", "T__10", "T__11", "T__12", 
+                      "T__13", "T__14", "T__15", "T__16", "T__17", "T__18", 
+                      "T__19", "T__20", "T__21", "T__22", "T__23", "T__24", 
+                      "T__25", "T__26", "T__27", "T__28", "T__29", "T__30", 
+                      "ABSTRACT", "ACTION", "FUNCTION", "KEY", "NULLABLE", 
+                      "NUMBER", "INT", "EXP", "STRING", "ESC", "UNICODE", 
+                      "HEX", "SAFECODEPOINT", "EDM", "ID", "QID", "TID", 
+                      "SIMPLE", "DOC_COMMENT", "LINE_COMMENT", "WS" ];
+
+    constructor(input) {
+        super(input)
+        this._interp = new antlr4.atn.LexerATNSimulator(this, atn, decisionsToDFA, new antlr4.PredictionContextCache());
+    }
+
+    get atn() {
+        return atn;
+    }
 }
-
-rsdlLexer.prototype = Object.create(antlr4.Lexer.prototype);
-rsdlLexer.prototype.constructor = rsdlLexer;
-
-Object.defineProperty(rsdlLexer.prototype, "atn", {
-        get : function() {
-                return atn;
-        }
-});
 
 rsdlLexer.EOF = antlr4.Token.EOF;
 rsdlLexer.T__0 = 1;
@@ -341,45 +366,5 @@ rsdlLexer.DOC_COMMENT = 43;
 rsdlLexer.LINE_COMMENT = 44;
 rsdlLexer.WS = 45;
 
-rsdlLexer.prototype.channelNames = [ "DEFAULT_TOKEN_CHANNEL", "HIDDEN" ];
 
-rsdlLexer.prototype.modeNames = [ "DEFAULT_MODE" ];
-
-rsdlLexer.prototype.literalNames = [ null, "'namespace'", "'include'", "'as'", 
-                                     "'type'", "'{'", "'}'", "'extends'", 
-                                     "':'", "'['", "']'", "'Boolean'", "'Date'", 
-                                     "'DateTime'", "'Decimal'", "'('", "','", 
-                                     "')'", "'Double'", "'Duration'", "'Integer'", 
-                                     "'String'", "'TimeOfDay'", "'enum'", 
-                                     "'flags'", "'typedef'", "'service'", 
-                                     "'true'", "'false'", "'null'", "'.'", 
-                                     "'/'", "'abstract'", "'action'", "'function'", 
-                                     "'key'", "'?'" ];
-
-rsdlLexer.prototype.symbolicNames = [ null, null, null, null, null, null, 
-                                      null, null, null, null, null, null, 
-                                      null, null, null, null, null, null, 
-                                      null, null, null, null, null, null, 
-                                      null, null, null, null, null, null, 
-                                      null, null, "ABSTRACT", "ACTION", 
-                                      "FUNCTION", "KEY", "NULLABLE", "NUMBER", 
-                                      "STRING", "EDM", "ID", "QID", "TID", 
-                                      "DOC_COMMENT", "LINE_COMMENT", "WS" ];
-
-rsdlLexer.prototype.ruleNames = [ "T__0", "T__1", "T__2", "T__3", "T__4", 
-                                  "T__5", "T__6", "T__7", "T__8", "T__9", 
-                                  "T__10", "T__11", "T__12", "T__13", "T__14", 
-                                  "T__15", "T__16", "T__17", "T__18", "T__19", 
-                                  "T__20", "T__21", "T__22", "T__23", "T__24", 
-                                  "T__25", "T__26", "T__27", "T__28", "T__29", 
-                                  "T__30", "ABSTRACT", "ACTION", "FUNCTION", 
-                                  "KEY", "NULLABLE", "NUMBER", "INT", "EXP", 
-                                  "STRING", "ESC", "UNICODE", "HEX", "SAFECODEPOINT", 
-                                  "EDM", "ID", "QID", "TID", "SIMPLE", "DOC_COMMENT", 
-                                  "LINE_COMMENT", "WS" ];
-
-rsdlLexer.prototype.grammarFileName = "rsdl.g4";
-
-
-exports.rsdlLexer = rsdlLexer;
 
