@@ -1,10 +1,12 @@
-const expect = require("chai").expect;
+import { expect } from "chai";
+import { exec } from "child_process";
+import fs from "fs";
+import path from "path";
 
-const exec = require("child_process").exec;
-const fs = require("fs");
-const path = require("path");
+import url from "url";
+const __dirname = url.fileURLToPath(new URL(".", import.meta.url));
 
-describe("CLI", () => {
+describe("rsdl-js CLI", () => {
   it("Help", async () => {
     const result = await cmd(["-h"]);
     expect(result.code).to.equal(0);
@@ -59,7 +61,7 @@ describe("CLI", () => {
   });
 });
 
-describe("CLI - error cases", () => {
+describe("rsdl-js CLI - error cases", () => {
   it("Invalid option", async () => {
     const result = await cmd(["-x"]);
     expect(result.code).to.equal(0);
