@@ -3,7 +3,9 @@ import { App, Listener } from "./app";
 import { csdl2openapi } from "odata-openapi/lib/csdl2openapi";
 import { parse } from "rsdl-js/lib/parser";
 import { serializeToXml } from "csdl2xml";
-
+import hljs from "highlight.js/lib/core";
+import json from "highlight.js/lib/languages/json";
+import xml from "highlight.js/lib/languages/xml";
 
 class ViewerListener extends Listener {
   private host: string;
@@ -16,7 +18,7 @@ class ViewerListener extends Listener {
   }
 
   callback = (rsdl: string) => {
-    
+
     // Get various forms of representation
     const csdlJson = convert2csdl(rsdl);
     const openapi = convertCsdlJson2OpenApi(csdlJson, {
@@ -29,7 +31,9 @@ class ViewerListener extends Listener {
     // set content for CSDL, CSDL-XML, and OpenAPI tabs
     this.config.openApiTabContent.innerHTML = JSON.stringify(openapi, null, 2);
     this.config.csdlTabContent.innerHTML = JSON.stringify(csdlJson, null, 2);
-    this.config.csdlXmlTabContent.innerText = xml;
+    this.config.csdlXmlTabContent.innerHTML = xml
+      .replace(/</g, "&lt;")
+      .replace(/>/g, "&gt;");
 
     // update SwaggerUI content
     SwaggerUI({
@@ -37,12 +41,16 @@ class ViewerListener extends Listener {
       spec: openapi,
       supportedSubmitMethods: [],
     });
+
+    hljs.highlightAll();
   };
 }
 
 function initViewers(app: App, host, basePath) {
   const viewerListener = new ViewerListener(host, basePath);
   app.addListener(viewerListener);
+  hljs.registerLanguage("json", json);
+  hljs.registerLanguage("xml", xml);
   return viewerListener;
 }
 
@@ -67,7 +75,7 @@ function convert2csdl(rsdl: string) {
 function convert2csdlxml(csdl) {
   try {
     return serializeToXml(csdl);
-   } catch (e) {
+  } catch (e) {
     console.error(e);
     return e;
   }
