@@ -1,97 +1,131 @@
-module.exports = {
-  title: 'RAPID',
-  tagline: 'Fastest way to build your RESTful API',
-  url: 'https://rapid.rocks',
-  baseUrl: '/',
-  favicon: 'img/rest.svg',
-  organizationName: 'oasis-open', // Usually your GitHub org/user name.
-  projectName: 'odata-rapid', // Usually your repo name.
-  staticDirectories: [
-    'static',
-    '../tools/api-designer/dist',
-    '../tools/api-explorer/dist',
-  ],
-  themeConfig: {
-    prism: {
-      theme: require('prism-react-renderer/themes/oceanicNext'),
-    },
-    navbar: {
-      title: 'RAPID',
-      logo: {
-        alt: 'RAPID Logo',
-        src: 'img/logo.svg',
-      },
-      items: [
-        {
-          to: 'docs/',
-          activeBasePath: 'docs',
-          label: 'Docs',
-          position: 'left',
-        },
-        {
-          to: 'docs/samples/apidesigner',
-          label: 'Samples',
-          position: 'left',
-          activeBaseRegex: `docs/(samples/apidesigner|samples/apiexplorer|samples/jetsons|tutorial/graphqlclient)`,
-        },
-      ],
-    },
-    footer: {
-      style: 'dark',
-      links: [
-        {
-          title: 'Docs',
-          items: [
-            {
-              label: 'Introduction',
-              to: 'docs/',
-            },
-          ],
-        },
-        {
-          title: 'Community',
-          items: [
-            {
-              label: 'Discord',
-              href: 'https://discord.gg/w9YPQAp',
-            },
-            {
-              label: 'Twitter',
-              href: 'https://twitter.com/rapidprofile',
-            },
-          ],
-        },
-        {
-          title: 'More',
-          items: [
-            {
-              label: 'GitHub',
-              href: 'https://github.com/oasis-open/odata-rapid',
-            },
-          ],
-        },
-      ],
-      copyright: `Copyright © ${new Date().getFullYear()} OASIS OPEN,`,
-    },
+// @ts-check
+// Note: type annotations allow type checking and IDEs autocompletion
+
+const lightCodeTheme = require("prism-react-renderer/themes/oceanicNext");
+const darkCodeTheme = require("prism-react-renderer/themes/dracula");
+
+/** @type {import('@docusaurus/types').Config} */
+const config = {
+  title: "RAPID",
+  tagline: "Fastest way to build your RESTful API",
+  url: "https://rapid.rocks",
+  baseUrl: "/",
+  onBrokenLinks: "throw",
+  onBrokenMarkdownLinks: "warn",
+  favicon: "img/rest.svg",
+
+  // GitHub pages deployment config.
+  // If you aren't using GitHub pages, you don't need these.
+  organizationName: "oasis-open", // Usually your GitHub org/user name.
+  projectName: "odata-rapid", // Usually your repo name.
+
+  // Even if you don't use internalization, you can use this field to set useful
+  // metadata like html lang. For example, if your site is Chinese, you may want
+  // to replace "en" with "zh-Hans".
+  i18n: {
+    defaultLocale: "en",
+    locales: ["en"],
   },
-  scripts: [
-    'https://cdnjs.cloudflare.com/ajax/libs/gsap/1.16.1/TweenMax.min.js',
-    'https://s3-us-west-2.amazonaws.com/s.cdpn.io/16327/MorphSVGPlugin.min.js',
-  ],
+
   presets: [
     [
-      '@docusaurus/preset-classic',
-      {
+      "classic",
+      /** @type {import('@docusaurus/preset-classic').Options} */
+      ({
         docs: {
-          path: '../docs',
-          routeBasePath: 'docs',
-          sidebarPath: require.resolve('./sidebars.js'),
-          editUrl: 'https://github.com/oasis-open/odata-rapid/edit/main/docs/',
+          path: "../docs",
+          sidebarPath: require.resolve("./sidebars.js"),
+          // Please change this to your repo.
+          // Remove this to remove the "edit this page" links.
+          editUrl: "https://github.com/oasis-open/odata-rapid/edit/main/docs/",
         },
         theme: {
-          customCss: require.resolve('./src/css/custom.css'),
+          customCss: require.resolve("./src/css/custom.css"),
         },
-      },
+      }),
     ],
   ],
+
+  staticDirectories: [
+    "static",
+    "../tools/api-designer/dist",
+    "../tools/api-explorer/dist",
+  ],
+
+  themeConfig:
+    /** @type {import('@docusaurus/preset-classic').ThemeConfig} */
+    ({
+      colorMode: {
+        defaultMode: "light",
+        disableSwitch: true,
+        respectPrefersColorScheme: false,
+      },
+      navbar: {
+        title: "RAPID",
+        logo: {
+          alt: "RAPID Logo",
+          src: "img/logo.svg",
+        },
+        items: [
+          {
+            to: "docs",
+            label: "Docs",
+            position: "left",
+          },
+          {
+            to: "docs/samples/apidesigner",
+            label: "Samples",
+            position: "left",
+          },
+          {
+            href: "https://github.com/oasis-open/odata-rapid",
+            label: "GitHub",
+            position: "right",
+          },
+        ],
+      },
+      footer: {
+        style: "dark",
+        links: [
+          {
+            title: "Docs",
+            items: [
+              {
+                label: "Introduction",
+                to: "docs",
+              },
+            ],
+          },
+          {
+            title: "Community",
+            items: [
+              {
+                label: "Discord",
+                href: "https://discord.gg/w9YPQAp",
+              },
+              {
+                label: "Twitter",
+                href: "https://twitter.com/rapidprofile",
+              },
+            ],
+          },
+          {
+            title: "More",
+            items: [
+              {
+                label: "GitHub",
+                href: "https://github.com/oasis-open/odata-rapid",
+              },
+            ],
+          },
+        ],
+        copyright: `Copyright © ${new Date().getFullYear()} OASIS OPEN. Built with Docusaurus.`,
+      },
+      prism: {
+        theme: lightCodeTheme,
+        darkTheme: darkCodeTheme,
+      },
+    }),
 };
+
+module.exports = config;
