@@ -1,50 +1,59 @@
-module.exports = {
-    docs: [
-        {
-            type: 'category',
-            label: 'Introduction',
-            items: [
-                'gettingstarted'
-            ]
-        },
-        {
-            type: 'category',
-            label: 'Features',
-            items: [
-                'spec/features',
-                'rapid-read',
-                'rapid-edit',
-                'operations',
-                'spec/servicedescription',
-                'spec/batch'
-            ]
-        },
-        {
-            type: 'category',
-            label: 'RAPID Schema Definition Language (RSDL)',
-            items: [
-                'rsdl/rsdl-intro',
-                'rsdl/rsdl-semantics',
-                'rsdl/rsdl-abnf',
-            ]
-        },
-        {
-            type: 'category',
-            label: 'Related Technologies',
-            items: [
-                'related/odata',
-                'related/graphql',
-                'related/openapi'
-            ]
-        }
+/**
+ * Creating a sidebar enables you to:
+ - create an ordered group of docs
+ - render a sidebar for each doc of that group
+ - provide next/previous navigation
+
+ The sidebars can be generated from the filesystem, or explicitly defined here.
+
+ Create as many sidebars as you want.
+ */
+
+// @ts-check
+
+/** @type {import('@docusaurus/plugin-content-docs').SidebarsConfig} */
+const sidebars = {
+  // By default, Docusaurus generates a sidebar from the docs folder structure
+  // tutorialSidebar: [{type: 'autogenerated', dirName: '.'}],
+
+  // But you can create a sidebar manually
+  docs: [
+    {
+      type: "category",
+      label: "Introduction",
+      items: ["gettingstarted"],
+    },
+    {
+      type: "category",
+      label: "Features",
+      items: [
+        "spec/features",
+        "rapid-read",
+        "rapid-edit",
+        "operations",
+        "spec/servicedescription",
+        "spec/batch",
+      ],
+    },
+    {
+      type: "category",
+      label: "RAPID Schema Definition Language (RSDL)",
+      items: ["rsdl/rsdl-intro", "rsdl/rsdl-semantics", "rsdl/rsdl-abnf"],
+    },
+    {
+      type: "category",
+      label: "Related Technologies",
+      items: ["related/odata", "related/graphql", "related/openapi"],
+    },
+  ],
+  samples: {
+    Samples: [
+      "samples/apidesigner",
+      "samples/apiexplorer",
+      "samples/jetsons",
+      "tutorial/graphqlclient",
     ],
-    samples: 
-        {
-           'Samples' : [
-               'samples/apidesigner',
-               'samples/apiexplorer',
-               'samples/jetsons',
-               'tutorial/graphqlclient'
-            ]
-       }
+  },
 };
+
+module.exports = sidebars;
