@@ -1,110 +1,43 @@
-import React from 'react';
-import classnames from 'classnames';
-import Layout from '@theme/Layout';
-import Link from '@docusaurus/Link';
-import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
-import useBaseUrl from '@docusaurus/useBaseUrl';
-import styles from './styles.module.css';
-// Temporary bug in docusaurus
-// import Rocket from '@site/static/img/rocket.svg';
-// import animateRapidSpaceMan from '../components/rocket/animation'
-import RabbitSvg from '@site/static/img/rabbit.svg';
+import React from "react";
+import clsx from "clsx";
+import Link from "@docusaurus/Link";
+import useDocusaurusContext from "@docusaurus/useDocusaurusContext";
+import Layout from "@theme/Layout";
+import HomepageFeatures from "@site/src/components/HomepageFeatures";
 
-const features = [
-  {
-    title: <>Easy to Use</>,
-    // imageUrl: 'img/undraw_docusaurus_mountain.svg',
-    description: (
-      <>
-        Rapid provides a fast and robust approach for building your REST-based
-        APIs
-      </>
-    ),
-  },
-  {
-    title: <>Focus on What Matters</>,
-    //  imageUrl: 'img/undraw_docusaurus_tree.svg',
-    description: (
-      <>
-        Rapid allows you to focus on your API's resource model and rely on
-        well-established conventions for the nitty-gritty details
-      </>
-    ),
-  },
-  {
-    title: <>Based on Open Standards</>,
-    //   imageUrl: 'img/undraw_docusaurus_react.svg',
-    description: (
-      <>
-        Rapid is an Open Specification built by OASIS members and the community.
-        Rapid is based on existing standards like OData, OpenAPI, and GraphQL
-      </>
-    ),
-  },
-];
+import styles from "./index.module.css";
 
-function Feature({ imageUrl, title, description }) {
-  const imgUrl = useBaseUrl(imageUrl);
+import RabbitSvg from "@site/static/img/rabbit.svg";
+
+function HomepageHeader() {
+  const { siteConfig } = useDocusaurusContext();
   return (
-    <div className={classnames('col col--4', styles.feature)}>
-      {imgUrl && (
-        <div className="text--center">
-          <img className={styles.featureImage} src={imgUrl} alt={title} />
+    <header className={clsx("hero hero--primary", styles.heroBanner)}>
+      <div className="container">
+        {/* <h1 className="hero__title">{siteConfig.title}</h1> */}
+        <RabbitSvg />
+        <p className="hero__subtitle">{siteConfig.tagline}</p>
+        <div className={styles.buttons}>
+          <Link className="button button--secondary button--lg" to="docs">
+            Get started
+          </Link>
         </div>
-      )}
-      <h3>{title}</h3>
-      <p>{description}</p>
-    </div>
+      </div>
+    </header>
   );
 }
 
-function Home() {
-  const context = useDocusaurusContext();
-  const { siteConfig = {} } = context;
+export default function Home() {
+  const { siteConfig } = useDocusaurusContext();
   return (
     <Layout
-      title={`${siteConfig.title}`}
+      title={`Hello from ${siteConfig.title}`}
       description="Description will go into a meta tag in <head />"
     >
-      <header className={classnames('hero hero--primary', styles.heroBanner)}>
-        <div className="container">
-          <div className="row">
-            <div className={classnames('col')}>
-              {/* <h1 className="hero__title">{siteConfig.title}</h1> */}
-              <RabbitSvg />
-              <p className="hero__subtitle">{siteConfig.tagline}</p>
-              <div className={styles.buttons}>
-                <Link
-                  className={classnames(
-                    'button button--outline button--secondary button--lg',
-                    styles.getStarted,
-                  )}
-                  to={useBaseUrl('docs/')}
-                >
-                  Get Started
-                </Link>
-              </div>
-            </div>
-          </div>
-        </div>
-      </header>
+      <HomepageHeader />
       <main>
-        {features && features.length > 0 && (
-          <section className={styles.features}>
-            <div className="container">
-              <div className="row">
-                {features.map((props, idx) => (
-                  <Feature key={idx} {...props} />
-                ))}
-              </div>
-            </div>
-          </section>
-        )}
+        <HomepageFeatures />
       </main>
     </Layout>
   );
 }
-
-// setTimeout(animateRapidSpaceMan, 1000)
-
-export default Home;
